Made some progress.
Set up delete button. It is dangerous and I need to set it to have a confirmation window, but I'll do that later. I need to
fix the issues with the date again, which I'm not very excited about, but alas! I think it's in the creation part, and it only affects
the database. That's all I know. Maybe the reading in of the shirts on startup is the issue.