Fixed date issue with pick(), improved confirmation.
It was using getSeconds(), not getTime().
I made it so that whe you pick a shirt, the shirt blinks on the screen for a minute
so you know you did something. Next I'm going to set up a long-tap
menu so you can delete and stuff directly from the home screen.